Refactor part1, add modules, tests, and cleanup
