+++ part1/__init__.py
@@ -1,0 +1,14 @@
+from .gaussian import gaussian_eliminate, back_substitution
+from .determinant import determinant
+from .inverse import inverse
+from .rank_basis import rank_and_basis
+from .verify import verify_solution
+
+__all__ = [
+    "gaussian_eliminate",
+    "back_substitution",
+    "determinant",
+    "inverse",
+    "rank_and_basis",
+    "verify_solution",
+]
